refactor: extract email templates into separate files
