+++ api/src/lib/email/template/response.ts
@@ -3,39 +3,7 @@ import { Event, PublicEvent, Response } from 'types/graphql'
 
 import { SITE_HOST } from 'src/app.config'
 
-import {
-  prettyEndWithBetween,
-  prettyStartWithUntil,
-  prettyUntil,
-} from '../convert/date'
-
-import { sendEmail } from './send'
-
-/**
- * Email the admin link for a newly-created event to its owner.
- * @param event The newly-created event
- * @returns The result of the send operation
- */
-export async function sendEventDetails(
-  event: Pick<Event, 'ownerEmail' | 'title' | 'editToken'>
-) {
-  return sendEmail({
-    to: event.ownerEmail,
-    subject: `Manage your event: ${event.title}`,
-    text: stripIndent`
-			Hello from Freevite! Click this link to manage your event details and make it public:
-
-			https://${SITE_HOST}/edit?token=${event.editToken}
-
-			You must click the above link within 24 hours to confirm your email address.
-			Otherwise, we will automatically delete your event. Feel free to recreate it.
-
-			If you did not create this event, you can ignore this email and this event will be deleted.
-
-			If you need any help, just reply to this email. Thanks for using Freevite!
-		`,
-  })
-}
+import { sendEmail } from '../send'
 
 /**
  * Send a "confirm RSVP" email to the user who just responded to an event.
@@ -98,41 +66,6 @@ export async function sendNewResponseReceived({
       https://${SITE_HOST}/edit?token=${event.editToken}
 
       If you need any help, just reply to this email. Thanks for using Freevite!
-    `,
-  })
-}
-
-/**
- * Remind a user about an upcoming event for which they RSVPed.
- * @param event The event for which the reminder is being sent
- * @param response The response to which the reminder belongs
- * @param now The current date and time
- * @returns The result of the send operation
- */
-export async function sendReminder({
-  event,
-  response,
-  now,
-}: {
-  event: Pick<PublicEvent, 'title' | 'start' | 'end' | 'timezone' | 'slug'>
-  response: Pick<Response, 'email'>
-  now: Date
-}) {
-  return sendEmail({
-    to: response.email,
-    subject: `Reminder: ${event.title} (${prettyUntil(event.start, now, false)} away)`,
-    text: stripIndent`
-      Hello from Freevite! You asked for a reminder about this event when you RSVPed:
-
-      ${event.title}
-      Starts at: ${prettyStartWithUntil(event.start, event.timezone, now)})
-      Ends at: ${prettyEndWithBetween(event.start, event.end, event.timezone)}
-
-      View the event here:
-      https://${SITE_HOST}/events/${event.slug}
-
-      If you need to modify or delete your RSVP, just reply to this email.
-      Thanks for using Freevite!
     `,
   })
 }
