+++ api/src/services/events/events.ts
@@ -7,7 +7,7 @@ import type {
 import { validate } from '@redwoodjs/api'
 
 import { db } from 'src/lib/db'
-import { sendEventDetails } from 'src/lib/email/template'
+import { sendEventDetails } from 'src/lib/email/template/event'
 import { notifyEventCreated, notifyEventUpdated } from 'src/lib/notify'
 import { summarize } from 'src/lib/response'
 import { generateToken, alphaLower } from 'src/lib/token'
